Move HttpHeadersImpl class to internal and fix #503
