+++ core/src/main/java/org/wso2/msf4j/internal/HttpHeadersImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.msf4j;
+package org.wso2.msf4j.internal;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
